Extract code to return a login result to LoginResultHandler
