+++ pkg/auth/handler/saml/deps.go
@@ -7,6 +7,7 @@ import (
 var DependencySet = wire.NewSet(
 	NewLoginHandlerLogger,
 	NewLoginFinishHandlerLogger,
+	wire.Struct(new(LoginResultHandler), "*"),
 	wire.Struct(new(MetadataHandler), "*"),
 	wire.Struct(new(LoginHandler), "*"),
 	wire.Struct(new(LoginFinishHandler), "*"),
